fix: Corriger le positionnement du floor et refactoriser les systèmes ECS
- Corriger SPRITE_HEIGHT de 208.0f à 86.0f pour correspondre à la taille réelle de la texture
- Supprimer FLOOR_OFFSET_Y pour coller le floor au bas de l'écran
- Modifier setSpriteScale pour accepter float au lieu de int afin de supporter les scales < 1.0
- Appliquer le scale uniquement pour les sprites avec scrolling (mis à jour dynamiquement)
- Les autres sprites conservent leur scale initial défini à la création
- Refactoriser les systèmes ECS génériques (Projectile, Scrolling, Explosion, PlayerDirection, LoadingAnimation, Animation, Debug)
- Déplacer les systèmes dupliqués vers modules/ECS/Systems pour une meilleure réutilisabilité

+++ modules/Interfaces/include/Interfaces/IRenderer.hpp
@@ -147,7 +147,7 @@ namespace eng
             virtual void drawSprite(const std::string &name) = 0;
             virtual void setSpritePosition(const std::string &name, float x, float y) = 0;
             virtual void setSpriteTexture(const std::string &name, const std::string &path) = 0;
-            virtual void setSpriteScale(const std::string &name, int x, int y) = 0;
+            virtual void setSpriteScale(const std::string &name, float x, float y) = 0;
             virtual void setSpriteFrame(const std::string &name, int fx, int fy, int fnx, int fny) = 0;
             virtual void setSpriteRotation(const std::string &name, float rotation) = 0;
             virtual void rotateSprite(const std::string &name, float rotation) = 0;
