+++ plugins/Renderer/SFML/include/SFMLRenderer/SFMLRenderer.hpp
@@ -61,7 +61,7 @@ namespace eng
             void setSpritePosition(const std::string &name, float x, float y) override;
             void setSpriteTexture(const std::string &name, const std::string &path) override;
             void setSpriteFrame(const std::string &name, int fx, int fy, int fnx, int fny) override;
-            void setSpriteScale(const std::string &name, int x, int y) override;
+            void setSpriteScale(const std::string &name, float x, float y) override;
             void setSpriteColor(const std::string &name, Color color) override;
             void drawSprite(const std::string &name) override;
             void setSpriteRotation(const std::string &name, float rotation) override;
